Add v2 features: CloudKit, widget, merchant DB, PDF export, Siri, Spotlight
- CloudKit sync (SwiftData cloudKitDatabase: .automatic)
- VNDocumentCameraViewController for better OCR quality
- NaturalLanguage NER for merchant name extraction
- Merchant database (150+ merchants) with fuzzy matching
- PDF export service with monthly reports
- WidgetKit spending widget (small + medium sizes)
- App Intents for Siri ("How much did I spend today/this week?")
- CoreSpotlight indexing for receipt search
- Export tab with month picker and category breakdown
- Share receipts with photo + text via UIActivityViewController
- Updated tests for merchant DB integration

+++ ReceiptBrain/Views/DashboardView.swift
@@ -57,7 +57,7 @@ struct DashboardView: View {
             } else {
                 Chart(data, id: \.category) { item in
                     SectorMark(
-                        angle: .value("Amount", item.total as NSDecimalNumber),
+                        angle: .value("Amount", (item.total as NSDecimalNumber).doubleValue),
                         innerRadius: .ratio(0.5)
                     )
                     .foregroundStyle(by: .value("Category", item.category.displayName))
@@ -85,7 +85,7 @@ struct DashboardView: View {
                 Chart(data, id: \.weekStart) { item in
                     BarMark(
                         x: .value("Week", item.weekStart, unit: .weekOfYear),
-                        y: .value("Amount", item.total as NSDecimalNumber)
+                        y: .value("Amount", (item.total as NSDecimalNumber).doubleValue)
                     )
                     .foregroundStyle(.blue.gradient)
                 }
